refactor: move to command
