+++ src/features/MainScreen/NotesListPanel/NotesList.tsx
@@ -7,7 +7,6 @@ import { getContextMenuCoords } from '@electron/requests/contextMenu/renderer';
 import { useNoteContextMenu } from '@features/NotesContainer/NoteContextMenu/useNoteContextMenu';
 import { useTelemetryTracker } from '@features/telemetry';
 import { useNoteActions } from '@hooks/notes/useNoteActions';
-import { useUpdateNotes } from '@hooks/notes/useUpdateNotes';
 import { useIsActiveWorkspace } from '@hooks/useIsActiveWorkspace';
 import { useWorkspaceSelector } from '@state/redux/profiles/hooks';
 import {
@@ -27,7 +26,6 @@ export type NotesListProps = {};
 export const NotesList: FC<NotesListProps> = () => {
 	const telemetry = useTelemetryTracker();
 
-	const updateNotes = useUpdateNotes();
 	const noteActions = useNoteActions();
 
 	const activeNoteId = useWorkspaceSelector(selectActiveNoteId);
@@ -35,9 +33,7 @@ export const NotesList: FC<NotesListProps> = () => {
 
 	const search = useWorkspaceSelector(selectSearch);
 
-	const openNoteContextMenu = useNoteContextMenu({
-		updateNotes,
-	});
+	const openNoteContextMenu = useNoteContextMenu();
 
 	const parentRef = useRef<HTMLDivElement>(null);
 	const isActiveWorkspace = useIsActiveWorkspace();
