+++ src/features/NotesContainer/OpenedNotesPanel.tsx
@@ -25,11 +25,8 @@ export const OpenedNotesPanel: FC<TopBarProps> = ({
 	activeTab,
 	onClose,
 	onPick,
-	updateNotes,
 }) => {
-	const openNoteContextMenu = useNoteContextMenu({
-		updateNotes,
-	});
+	const openNoteContextMenu = useNoteContextMenu();
 
 	const existsTabs = useMemo(
 		() => tabs.filter((noteId) => notes.some((note) => note.id === noteId)),
