+++ src/hooks/commands/index.ts
@@ -53,6 +53,11 @@ export enum GLOBAL_COMMANDS {
 	COPY_NOTE_MARKDOWN_LINK = 'copyNoteMarkdownLink',
 
 	/**
+	 * Creates a copy of an existing note
+	 */
+	DUPLICATE_NOTE = 'DUPLICATE_NOTE',
+
+	/**
 	 * Switch focus to the next open note
 	 */
 	FOCUS_NEXT_NOTE = 'focusNextNote',
@@ -98,6 +103,7 @@ type CommandsWithPayload = {
 
 	[GLOBAL_COMMANDS.EXPORT_NOTE]: { noteId: string };
 	[GLOBAL_COMMANDS.COPY_NOTE_MARKDOWN_LINK]: { noteId: string };
+	[GLOBAL_COMMANDS.DUPLICATE_NOTE]: { noteId: string };
 
 	[GLOBAL_COMMANDS.SHORTCUTS_PRESSED]: Shortcuts;
 };
